requirements.txt
Django==3.2.10
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
drf-yasg==1.20.0
gunicorn==20.1.0
whitenoise==6.7.0